Add ssh-agent start on git-bash start
The ssh-agent is not starting when Windows start, so I need to keep
entering the ssh-key passphrase every time a git command is run against
a remote repository.

This change checks if the ssh-agent is started when the git bash
terminal starts, and if it is not, it will be started.